Invert `read_directory` and `add_watch` order
When copying a new page with a lot of content, some keys are missing.
This is because as soon as the directory page is copied, we read its
directory, but not all content may already be in the new directory. And
we start watching the directory once read, so some files can be missing
because "added" in the directory between both actions.

This is now fixed because we start watching the directory before reading
its content. A file may be seen twice but in this case it will just, for
example, rerender the key.

Fix #36